+++ hosts/chi/darwin-configuration.nix
@@ -3,6 +3,11 @@
 {
   networking.knownNetworkServices = [ "Ethernet" "Wi-Fi" ];
 
+  # Start garbage collection when less than 5 GiB free and stop once 15 GiB is free
+  nix.settings.auto-optimise-store = true;
+  nix.settings.min-free = 5 * 1024 * 1024 * 1024;
+  nix.settings.max-free = 15 * 1024 * 1024 * 1024;
+
   launchd.user.agents.echo = {
     serviceConfig.ProgramArguments =
       [ "${pkgs.utm}/bin/utmctl" "start" "echo" ];